backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clerk-backend-api>=4.2.0",
    "fastapi>=0.124.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "svix>=1.83.0",
    "uvicorn>=0.38.0",
]